+++ lib/haku/result.rb
@@ -4,13 +4,19 @@ require "active_support/string_inquirer"
 
 module Haku
   class Result
-    def initialize(status, response)
+    def initialize(status, payload)
       @status = ActiveSupport::StringInquirer.new(status.to_s)
-      @response = response
+      @payload = payload
 
-      @response.each_key do |key|
-        define_singleton_method(key) { @response[key] }
+      return unless @payload.respond_to?(:to_h)
+
+      @payload.to_h.each_key do |key|
+        define_singleton_method(key) { @payload[key] }
       end
+    end
+
+    def result
+      @payload
     end
 
     def success?